--- src/GeekBrains/Lesson8/Homework/Logic.java
@@ -1,194 +1,0 @@
-package GeekBrains.Lesson8.Homework;
-
-import java.util.Random;
-import java.util.Scanner;
-
-public class Logic {
-    static int SIZE;
-    static int DOTS_TO_WIN;
-
-    static final char DOT_X = 'X';
-    static final char DOT_O = 'O';
-    static final char DOT_EMPTY = '.';
-
-    static char[][] map;
-
-    static Random random = new Random();
-
-    static boolean gameFinished;
-
-    public static void go() {
-        gameFinished = true;
-
-        printMap();
-        if (checkWinLines(DOT_X, DOTS_TO_WIN)) {
-            System.out.println("Вы выиграли!!!");
-            return;
-        }
-        if (isFull()) {
-            System.out.println("Ничья");
-            return;
-        }
-
-        aiTurn();
-        printMap();
-        if (checkWinLines(DOT_O, DOTS_TO_WIN)) {
-            System.out.println("Комьютер победил");
-            return;
-        }
-        if (isFull()) {
-            System.out.println("Ничья");
-            return;
-        }
-
-        gameFinished = false;
-    }
-
-    static void initMap() {
-        map = new char[SIZE][SIZE];
-        for (int i = 0; i < SIZE; i++) {
-            for (int j = 0; j < SIZE; j++) {
-                map[i][j] = DOT_EMPTY;
-            }
-        }
-    }
-
-    static void printMap() {
-        System.out.print("  ");
-        for (int i = 0; i < SIZE; i++) {
-            System.out.print(i + 1 + " ");
-        }
-        System.out.println();
-        for (int i = 0; i < SIZE; i++) {
-            System.out.print(i + 1 + " ");
-            for (int j = 0; j < SIZE; j++) {
-                System.out.printf("%c ", map[i][j]);
-            }
-            System.out.println();
-        }
-    }
-
-    static void humanTurn(int x, int y) {
-        if (isCellValid(y, x)) {
-            map[y][x] = DOT_X;
-            go();
-        }
-    }
-
-    static void aiTurn() {
-        int x;
-        int y;
-
-        // Попытка победить самому
-        for (int i = 0; i < SIZE; i++) {
-            for (int j = 0; j < SIZE; j++) {
-                if (isCellValid(i, j)) {
-                    map[i][j] = DOT_O;
-                    if (checkWinLines(DOT_O, DOTS_TO_WIN)) {
-                        return;
-                    }
-                    map[i][j] = DOT_EMPTY;
-                }
-            }
-        }
-// Сбить победную линии противника, если осталось 1 ход для победы
-        for (int i = 0; i < SIZE; i++) {
-            for (int j = 0; j < SIZE; j++) {
-                if (isCellValid(i, j)) {
-                    map[i][j] = DOT_X;
-                    if (checkWinLines(DOT_X, DOTS_TO_WIN)) {
-                        map[i][j] = DOT_O;
-                        return;
-                    }
-                    map[i][j] = DOT_EMPTY;
-                }
-            }
-        }
-
-// Сбить победную линии противника, если осталось 2 хода для победы
-        for (int i = 0; i < SIZE; i++) {
-            for (int j = 0; j < SIZE; j++) {
-                if (isCellValid(i, j)) {
-                    map[i][j] = DOT_X;
-                    if (checkWinLines(DOT_X, DOTS_TO_WIN - 1) &&
-                            Math.random() < 0.5) { //  фактор случайности, чтобы сбивал не все время первый попавшийся путь.
-                        map[i][j] = DOT_O;
-                        return;
-                    }
-                    map[i][j] = DOT_EMPTY;
-                }
-            }
-        }
-
-// Сходить в произвольную не занятую ячейку
-
-        do {
-            x = random.nextInt(SIZE);
-            y = random.nextInt(SIZE);
-        } while (!isCellValid(y, x));
-        map[y][x] = DOT_O;
-    }
-
-
-    static boolean isCellValid(int y, int x) {
-        if (y < 0 || x < 0 || y >= SIZE || x >= SIZE) {
-            return false;
-        }
-        return map[y][x] == DOT_EMPTY;
-    }
-
-    static boolean isFull() {
-        for (int i = 0; i < SIZE; i++) {
-            for (int j = 0; j < SIZE; j++) {
-                if (map[i][j] == DOT_EMPTY) {
-                    return false;
-                }
-            }
-        }
-        return true;
-    }
-
-//    static boolean checkWin(char c) {
-//        if (map[0][0] == c && map[0][1] == c && map[0][2] == c) { return true; }
-//        if (map[1][0] == c && map[1][1] == c && map[1][2] == c) { return true; }
-//        if (map[2][0] == c && map[2][1] == c && map[2][2] == c) { return true; }
-//
-//        if (map[0][0] == c && map[1][0] == c && map[2][0] == c) { return true; }
-//        if (map[0][1] == c && map[1][1] == c && map[2][1] == c) { return true; }
-//        if (map[0][2] == c && map[1][2] == c && map[2][2] == c) { return true; }
-//
-//        if (map[0][0] == c && map[1][1] == c && map[2][2] == c) { return true; }
-//        if (map[0][2] == c && map[1][1] == c && map[2][0] == c) { return true; }
-//
-//        return false;
-//    }
-
-    static boolean checkLine(int cy, int cx, int vy, int vx, char dot, int dotsToWin) {
-        if (cx + vx * (dotsToWin - 1) > SIZE - 1 || cy + vy * (dotsToWin - 1) > SIZE - 1 ||
-                cy + vy * (dotsToWin - 1) < 0) {
-            return false;
-        }
-
-        for (int i = 0; i < dotsToWin; i++) {
-            if (map[cy + i * vy][cx + i * vx] != dot) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    static boolean checkWinLines(char dot, int dotsToWin) {
-        for (int i = 0; i < SIZE; i++) {
-            for (int j = 0; j < SIZE; j++) {
-                if (checkLine(i, j, 0, 1, dot, dotsToWin) ||
-                        checkLine(i, j, 1, 0, dot, dotsToWin) ||
-                        checkLine(i, j, 1, 1, dot, dotsToWin) ||
-                        checkLine(i, j, -1, 1, dot, dotsToWin)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-}
