--- src/GeekBrains/Lesson8/Homework/TicTacToe.java
@@ -1,7 +1,0 @@
-package GeekBrains.Lesson8.Homework;
-
-public class TicTacToe {
-    public static void main(String[] args) {
-        new GameWindow();
-    }
-}
